Crate basehangul

An implementation of BaseHangul in Rust.

Structs

Decoder
An iterator adapter that decodes the byte stream from the BaseHangul stream.

Encoder
An iterator adapter that encodes the byte stream into the BaseHangul stream.

Packer
An iterator adapter that packs the byte stream into a stream of unsigned integers < 2,350.

Unpacker
An iterator adapter that unpacks the byte stream from a stream of unsigned integers < 2,350.

Functions

decode
Converts a BaseHangul string into the byte sequence if possible.

encode
Converts a byte sequence into the BaseHangul string.

Type Definitions

DecodeResult
The enum that contains either T or a decoding error.